crispy-forms-govuk/pyproject.toml
[metadata]
name = "crispy_forms_govuk"
version = "0.1.0"

[tool.poetry]
name = "crispy-forms-govuk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django = "~3.0.0"
django-crispy-forms = "~1.9.0"

[tool.poetry.dev-dependencies]
beautifulsoup4 = "^4.9.3"
black = "^21.6b0"
django-debug-toolbar = "^2.2.0"
doc8 = "^0.8"
docformatter = "^1.4"
flake8 = "^3.8.4"
flake8-pytest-style = "^1.3"
ipdb = "^0.13.4"
ipython = "^7.20.0"
isort = "^5.7.0"
lxml = "^4.6.0"
m2r2 = "^0.2"
mypy = "^0.800"
pytest = "^6.2"
pytest-cov = "^2.11"
pytest-django = "^4.4.0"
pytest-sugar = "^0.9.0"
sphinx = "^3.5"
sphinx-autobuild = "^2021.3.14"
sphinx-autodoc-typehints = "^1.11"
sphinx-rtd-theme = "^0.5.2"
tomlkit = "^0.7"

[build-system]
requires = ["setuptools >= 40.6.0", "wheel"]
build-backend = "setuptools.build_meta"
